Move Heiman specific converters to "local". Improve typing for Heiman specific attributes and make use of the cluster definitions.
